+++ odoo/local-src/sf_stock/__manifest__.py
@@ -8,9 +8,13 @@
     'category': 'SenseFly',
     'depends': ['stock', 'sale_stock'],
     'data': [
+        'security/ir.model.access.csv',
         'views/stock_location.xml',
         'views/stock_production_lot.xml',
         'views/stock_picking.xml',
+        'views/stock_inventory_category.xml',
+        'views/stock_inventory.xml',
+        'views/product_inventory_category.xml',
         'report/sf_report_deliveryslip.xml',
         'report/sf_report_delivery_note.xml',
         'report/sf_stock_report.xml'