+++ odoo/local-src/sf_stock/models/__init__.py
@@ -4,3 +4,6 @@ import stock_location
 from . import stock_production_lot
 from . import stock_move
 from . import stock_pack_operation
+from . import stock_inventory_category
+from . import stock_inventory
+from . import product_inventory_category