+++ HISTORY.rst
@@ -75,6 +75,7 @@ latest (unreleased)
 * Add draft image to be used over the reports
 * Add link beetween DO and Crystal report packing list
 * Quotation / Order report
+* Add stock inventory category filter on inventory adjustments
 * Add carrier accounts on partner
 * Add secondary vendor field on purchase order
 * Add sale order main partner to invoice email template